move attach/detach MediaStream methods into webrtc proxy
