--- packages/common/src/webrtc/utils.native.ts
@@ -1,7 +1,0 @@
-const attachMediaStream = (htmlElementId: string, stream: MediaStream) => null
-const detachMediaStream = (htmlElementId: string) => null
-
-export {
-  attachMediaStream,
-  detachMediaStream
-}
